Add several tests
