+++ app/tests/archives_tests/test_forms.py
@@ -263,44 +263,6 @@ def test_archive_item_form_unique_title(form_class):
 
 
 @pytest.mark.django_db
-def test_archive_item_detail_permissions(client):
-    archive = ArchiveFactory()
-    archive.add_editor(editor := UserFactory())
-    archive.add_uploader(uploader := UserFactory())
-
-    ai = ArchiveItemFactory(archive=archive)
-
-    def get_view(_user):
-        return get_view_for_user(
-            viewname="archives:item-detail",
-            client=client,
-            method=client.get,
-            reverse_kwargs={
-                "slug": archive.slug,
-                "pk": ai.pk,
-            },
-            user=_user,
-        )
-
-    # As uploader and uploader one can view
-    for user in uploader, editor:
-        response = get_view(user)
-        assert response.status_code == 200
-
-    # Removing the roles works
-    archive.remove_editor(editor)
-    archive.remove_uploader(uploader)
-
-    for user in (
-        editor,
-        uploader,
-        UserFactory(),  # Random user cannot view archive item
-    ):
-        response = get_view(user)
-        assert response.status_code == 403
-
-
-@pytest.mark.django_db
 def test_archive_item_update_permissions(client):
     archive = ArchiveFactory()
