+++ app/tests/archives_tests/test_models.py
@@ -1,7 +1,6 @@
 import pytest
 from django.db import IntegrityError, transaction
 
-from tests.algorithms_tests.factories import AlgorithmJobFactory
 from tests.archives_tests.factories import ArchiveFactory, ArchiveItemFactory
 from tests.components_tests.factories import ComponentInterfaceValueFactory
 
@@ -17,39 +16,6 @@ def create_archive_items_for_images(images, archive):
         civ = ComponentInterfaceValueFactory(image=image)
         ai = ArchiveItemFactory(archive=archive)
         ai.values.add(civ)
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize(
-    "distractor_generator",
-    (
-        # Different CIVs in distractor
-        lambda civs: [ComponentInterfaceValueFactory()],
-        # Partial CIVs in distractor
-        lambda civs: [civs[0]],
-        # Extra CIVs in distractor
-        lambda civs: [*civs, ComponentInterfaceValueFactory()],
-    ),
-)
-def test_archive_item_algorithm_job_as_input(distractor_generator):
-    ai = ArchiveItemFactory()
-    ai.values.set(
-        [
-            ComponentInterfaceValueFactory(),
-            ComponentInterfaceValueFactory(),
-        ]
-    )
-
-    aj1 = AlgorithmJobFactory()
-    aj2 = AlgorithmJobFactory()
-    for aj in aj1, aj2:
-        aj.inputs.set(ai.values.all())
-
-    distractor = AlgorithmJobFactory()
-    distractor_civs = distractor_generator(civs=ai.values.all())
-    distractor.inputs.set(distractor_civs)
-
-    assert {aj1.pk, aj2.pk} == {job.pk for job in ai.algorithm_jobs_as_input}
 
 
 @pytest.fixture(scope="function")
